feat: Puppeteer로 리액트 상태 추적 함수 감싸는 로직 추가
